test: extract policy white-box suite
