+++ scripts/rust-source-size-baseline.txt
@@ -3,7 +3,6 @@
 crates/agent-engine/skills/swebench/tooling/src/swebench_lite.rs 1389
 crates/agent-engine/src/agent_machine.rs 3661
 crates/agent-engine/src/config.rs 2556
-crates/agent-engine/src/policy.rs 1177
 crates/agent-engine/src/rollout.rs 2266
 crates/agent-engine/src/runtime/engine.rs 2816
 crates/agent-engine/src/runtime/submission_handlers.rs 2981
